% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEcutoff.R
\name{ATEcutoff}
\alias{ATEcutoff}
\title{Create ATE Cutoff}
\usage{
ATEcutoff(dataframe, formula, similarity_measures, type_model, k, bounds,
  user_seed, plot_treatment = NULL, plot_interact_x = NULL, stable_x)
}
\arguments{
\item{dataframe}{Dataframe from which we will estimate our regression model.}

\item{formula}{Symbolic representation of the model to be estimated. This is written in "typical" R language (i.e. y ~ x1 + x2), such that y is the outcome variable and x1 and x2 are the predictors.}

\item{similarity_measures}{Vector(s) from dataframe that contains the similarity measures to be used as weights. Possible values for measure_type = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"). Default is "jaccard".}

\item{type_model}{Statistical model to estimate. Currently support OLS and logistic ("ls", "logit").}

\item{k}{The penalty that you want to set for down-weighting inattentive respondents. Lower levels of k down-weight low attention participants more severely.}

\item{bounds}{Minimum and maximum of uniform distribution we should draw cutoff values between.}
}
\value{
Dataframe of all estimated marginal effects that make up the sampling distribution for participants that did and did not receive the treatment.
}
\description{
Create cutoff, then bootstrap or simulate the sampling distribution of participants that likely received the treatment (ATE among "compliers" and "non-compliers").
}
\examples{
ATEcutoff(dataframe=replication_complete.cases, similarity_measures=c("jaccardDist", "cosineDist"), bounds=0.2, formula=trustChurch_postTreat ~ Concordant*attendanceBin, model_type="ls", k=3))

}
\seealso{
\code{\link{plotComplierATE}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
