% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageSimilarity.R
\name{averageSimilarity}
\alias{averageSimilarity}
\title{Calculate Average Similarity}
\usage{
averageSimilarity(dataframe, similarity_measures, k)
}
\arguments{
\item{dataframe}{Dataframe from which you will select the similarity measures to be used to calculate the weights.}

\item{similarity_measures}{Vector(s) from dataframe that contains the similarity measures to be used as weights.}

\item{k}{The penalty that you want to set for down-weighting inattentive respondents. Lower levels of k down-weight low attention participants more severely.}
}
\value{
Original dataframe with an additional vector that is the averaged weight of the similarity measures provided.
}
\description{
Calculate the average weighted similarity for all participants.
}
\examples{
averageSimilarity(dataframe=replication_complete.cases, similarity_measures=c("jaccardDist", "cosineDist"),  k=3)

}
\seealso{
\code{\link{regressionComparison}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
