% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareK.R
\name{compareK}
\alias{compareK}
\title{Determine "best" level of k}
\usage{
compareK(dataframe, similarity_measures, correct_vec, k_range = 1:10,
  plot_path = NULL)
}
\arguments{
\item{dataframe}{Dataframe from which you will select the similarity measures from.}

\item{similarity_measures}{Vector(s) from dataframe that contains the similarity measures to be used as weights.}

\item{correct_vec}{Character indicating vector in dataframe that indicates whether a respondent answered "correctly" as determined by a human coder. Default is NULL, so if a user doesn't include anything function will automatically set a threshold of "correctness" based on a respondent's average similarity (i.e. those respondents that score below say 0.1 will not be included in the "list-wise deletion" sample.}

\item{k_range}{The range of penalties that you want plotted. Remember, lower levels of k down-weight low attention participants more severely.}

\item{plot_path}{If user wants to save figure, please provide a character vector for the file path in which the plot should be download. User must decide extension (pdf, jpg, png) in file path.}
}
\value{

}
\description{
Calculate the correlation between the "correct" answer as determined by a human and participants' average similarity score
}
\details{
Determine what value of k is best suited to reduce the impact of inattentive participants on the overall results, while still maintaining that our measure of attention is correlated with some indicator of correctness (even if it is subjective).
}
\examples{
compareK(dataframe=kaneData, 
  similarity_measures =c("jaccardSimilarity", "cosineSimilarity"), 
  correct_vec = "correct", 
  k_range=1:10)

}
\seealso{
\code{\link{similarityMeasures}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
