% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWordEmbeddings.R
\name{createWordEmbeddings}
\alias{createWordEmbeddings}
\title{Create Word Embeddings}
\usage{
createWordEmbeddings(df, prompts, responses, user_seed = 5,
  prune_len = 1)
}
\description{
Generate comparison between the prompt and participants' responses in a latent word2vec space.
}
\examples{


}
\seealso{
\code{\link{}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
