% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachAllPackages.R
\name{detachAllPackages}
\alias{detachAllPackages}
\title{detachAllPackages}
\usage{
detachAllPackages()
}
\description{
Clear global environment and detach all libraries
}
\details{
This function removes any hidden libraries or content that may be loaded in your global environment
}
\examples{

\dontrun{
detachAllPackages()
}
}
\seealso{
\code{\link{pkgTest}}
}
\author{
Jeffrey Ziegler
}
