% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeMarginalEffect.R
\name{executeMarginalEffect}
\alias{executeMarginalEffect}
\title{executeMarginalEffect}
\usage{
executeMarginalEffect(subset = NULL, regression_model)
}
\arguments{
\item{dataframe}{Dataframe that contains the prompt and response of the open-ended manipulation check.}

\item{regression_models}{Set of regression models that you wish to estimate the marginal effects for. In the example, there are five outcomes,  which appears in the example below.}
}
\value{
Dataframe of all estimated marginal effects that make up the sampling distribution.
}
\description{
Calculate the overall treatment effects differing how inattentive participants are down-weighted.
}
\examples{
executeMarginalEffect(subset="full", regression_model=baseModel_trustChurch_postTreat)

}
\seealso{
\code{\link{regressionComparison}} \code{\link{generateMarginalEffect}} \code{\link{plotMarginalEffect}} \code{\link{plotComplierATE}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
