% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMarginalEffect.R
\name{generateMarginalEffect}
\alias{generateMarginalEffect}
\title{Generate Marginal Effects}
\usage{
generateMarginalEffect(temp_model)
}
\arguments{
\item{temp_model}{Regression object indicating one model that you wish to calculate the marginal effects for.}
}
\value{
Dataframe of marginal effects.
}
\description{
This is an internal function that helps calculate the overall treatment effects differing how inattentive participants are down-weighted.
}
\examples{
generateMarginalEffect(baseModel_trustChurch_postTreat)

}
\seealso{
\code{\link{regressionComparison}} \code{\link{executeMarginalEffect}} \code{\link{plotMarginalEffect}} \code{\link{plotComplierATE}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
