% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVcomparison.R
\name{IVcomparison}
\alias{IVcomparison}
\title{Estimate Instrumental Variable Model for LATE}
\usage{
IVcomparison(dataframe = NULL, formula = NULL,
  similarity_measures = NULL, k = 3, up_down_weight = "up",
  print_regs = F)
}
\arguments{
\item{dataframe}{Dataframe from which you will select the outcome, predictors, and weights for the regression model.}

\item{formula}{Symbolic representation of the model to be estimated. This is written in "typical" R language (i.e. y ~ x1 + x2), such that y is the outcome variable and x_1,..., x_j are the predictors.}

\item{similarity_measures}{Vector(s) from dataframe that contains the similarity measures to be used as weights.}

\item{k}{The penalty that you want to set for down-weighting inattentive respondents. Lower levels of k down-weight low attention participants more severely.}

\item{up_down_weight}{Do you want to up-weight or down-weight?}

\item{print_regs}{Return table of estimated regression coefficients and fit statistics formatted for LaTeX using texreg(). Default=FALSE.}
}
\value{
Object containing instrumental variable regression call, data, results, etc.
}
\description{
Estimate the LATE using an instrumental variable approach to compare with the simulated sampling distribution of participants that likely received and didn't receive the treatment.
}
\details{
This function estimate an IV regression and generate a regression table of results with LaTeX formatting.
}
\seealso{
\code{\link{complierATE}} \code{\link{regressionComparison}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
