% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDependencies.R
\name{loadDependencies}
\alias{loadDependencies}
\title{loadDependencies}
\usage{
loadDependencies()
}
\arguments{
\item{packages}{A vector of package names that will be loaded as dependencies}
}
\description{
Clear global environment and detach all libraries
}
\details{
This function removes any hidden libraries or content that may be loaded in your global environment
}
\examples{

\dontrun{
loadDependencies()
}
}
\seealso{
\code{\link{pkgTest}}
}
\author{
Jeffrey Ziegler
}
