\name{openEnded-package}
\alias{openEnded-package}
\alias{openEnded}
\docType{package}
\title{
\packageTitle{openEnded}
}
\description{
\packageDescription{openEnded}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{openEnded}
\packageIndices{openEnded}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{openEnded}

Maintainer: \packageMaintainer{openEnded}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
