% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgTest.R
\name{pkgTest}
\alias{pkgTest}
\title{Check to see if a package is installed}
\usage{
pkgTest(pkg)
}
\description{
This function checks if a dependent library is installed, and if not installs it, and then loads it.
}
\examples{
lapply(c("stringdist", "ggplot2", "tidyr"), pkgTest)

}
\seealso{
\code{\link{loadDependencies}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
