% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrectness.R
\name{plotCorrectness}
\alias{plotCorrectness}
\title{Plot Similarity Measures and Objective Correctness}
\usage{
plotCorrectness(dataframe, measures, correct_vec, plot_path = NULL)
}
\arguments{
\item{dataframe}{Dataframe that contains the similarity measures you want to plot.}

\item{measures}{Two characters representing the vectors in dataframe that contain the similarity measures you want to compare.}

\item{correct_vec}{Character indicating vector in dataframe that indicates whether a respondent answered "correctly" as determined by a human coder. Must have this for plot to be used, Default=NULL.}
}
\description{
Plot respondents by similarity measure (user selects 2) and objective correctness.
}
\examples{



}
\seealso{
\code{\link{similarityMeasures}} \code{\link{plotSimilarity}}
}
\author{
Jeffrey Ziegler
}
