% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarginalEffect.R
\name{plotMarginalEffect}
\alias{plotMarginalEffect}
\title{plotMarginalEffect}
\usage{
plotMarginalEffect(regression_models)
}
\arguments{
\item{regression_models}{}

\item{dataframe}{Dataframe that contains the predictors and outcome.}
}
\value{
Plot of the marginal effects for the full, listwise, and weighted samples given the formula the user specified.
}
\description{
Plot the marginal effects varying which respondents are down-weighted in a regression.
}
\examples{
plotMarginalEffect(c(baseModel_trustChurch_postTreat, listwiseModel_trustChurch_postTreat, weightedModel_trustChurch_postTreat))
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
