% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimilarity.R
\name{plotSimilarity}
\alias{plotSimilarity}
\title{plotSimilarity}
\usage{
plotSimilarity(dataframe, measure, plot_path = NULL)
}
\arguments{
\item{dataframe}{Dataframe that contains the similarity measures you want to plot.}

\item{measure}{Name of vector in dataframe that contains the similarity measures you want to plot.}

\item{xlab}{Character/string of name user wants label of x-axis to be.}
}
\description{
This function plots the raw distribution for a given similarity measure.
}
\examples{

brazil_rows <- which(replication_complete.cases$Country=="Brazil")
plotSimilarity(dataframe=replication_complete.cases[brazil_rows,],
            measure="jaccardDist", xlab="Jaccard Similarity")

}
\seealso{
\code{\link{similarityMeasures}} \code{\link{plotSimilarityCorr}}
}
\author{
Jeffrey Ziegler
}
