% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimilarityCorr.R
\name{plotSimilarityCorr}
\alias{plotSimilarityCorr}
\title{plotSimilarityCorr}
\usage{
plotSimilarityCorr(dataframe, measures, labels)
}
\arguments{
\item{dataframe}{Dataframe that contains the similarity measures you want to plot.}

\item{measures}{Vector(s) in dataframe that contains the similarity measures you want to plot.}

\item{labels}{Character vector of labels for plot.}
}
\value{
The figure will appear in your plot window.
}
\description{
This function plots the correlation between a given similarity measure.
}
\examples{

plotSimilarityCorr(dataframe=replication_complete.cases, measures=c("jaccardDist", "cosineDist"))

}
\seealso{
\code{\link{similarityMeasures}}  \code{\link{plotSimilarity}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
