% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerCalc.R
\name{powerCalc}
\alias{powerCalc}
\title{Power Calculation}
\usage{
powerCalc(model, treatment_levels, x_levels, alpha = 0.05, side = "two-sided")
}
\arguments{
\item{model}{Model object from weighted regression that you wish to conduct a power analysis with.}

\item{treatment_levels}{How many treatment conditions are there?}

\item{x_levels}{How many levels does the factor variable that is interacted with the treatment have?}

\item{alpha}{Default = 0.05}

\item{side}{Default = "two-sided"}
}
\value{

}
\description{
Conduct a power calculation to make sure that you retain enough respondents to calculate a differential treatment effect if one likely exists.
}
\examples{
powerCalc()


}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
