% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepText.R
\name{prepText}
\alias{prepText}
\title{Clean and Tokenize Text}
\usage{
prepText(x)
}
\description{
Remove punctuation, capitalization, tokenize, etc. text prior to calculating distance in a word embedding space.
}
\examples{


}
\seealso{
\code{\link{}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
