% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightComparison.R
\name{regressionComparison}
\alias{regressionComparison}
\title{Compare the overall results varying on which respondents are included in the model based on attention}
\usage{
regressionComparison(dataframe, formula, model_type)
}
\arguments{
\item{dataframe}{Dataframe from which you will select the outcome, predictors, and weights for the regression model.}
}
\value{
Three regression objects, each one accessible as an element in a list
}
\description{
This function generates three regressions with:
(1) the full sample irrespective of attention
(2) a subsetted sample based on list-wise deletion of respondents that "failed" manipulation check
(3) weighted sample by forumal for average similarity
}
\examples{

\dontrun{
regressionComparison()
}
}
\author{
Jeffrey Ziegler
}
