% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityMeasures.R
\name{similarity_measures}
\alias{similarity_measures}
\title{simlarityMeasures}
\usage{
similarity_measures(
  dataframe,
  prompt,
  response,
  measure_type = "jaccard",
  ngrams
)
}
\arguments{
\item{#}{possible values for measure_type = c("osa", "lv", "dl", "hamming", "lcs",
"qgram", "cosine", "jaccard", "jw", "soundex")}
}
\description{
Clear global environment and detach all libraries
}
\details{
This function removes any hidden libraries or content that may be loaded in your global environment
}
\examples{

\dontrun{
simlarityMeasures()
}
}
\seealso{

}
\author{
Jeffrey Ziegler
}
