% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeMarginalEffect.R
\name{executeMarginalEffect}
\alias{executeMarginalEffect}
\title{Execute Marginal Effect}
\usage{
executeMarginalEffect(subset = NULL, regression_model = NULL)
}
\arguments{
\item{subset}{Which sample is used to estimate this regression model? Full ("base"), listwise deleting those that incorrectly answer manipulation check ("listwise"), or weighted using WLS based on continuous measure of attention ("weighted").}

\item{temp_model}{Regression object indicating one model that you wish to calculate the marginal effects for.}
}
\value{
Dataframe of all estimated marginal effects that make up the sampling distribution. If used with plotMarginalEffect, you will automatically receive the sampling distributions for all three types of subsets to compare against one another.
}
\description{
Run the essential generateMarginalEffects function to calculate the overall treatment effects differing how inattentive participants are down-weighted.
}
\examples{
executeMarginalEffect(subset="base", regression_model=baseModel_trustChurch_postTreat)

}
\seealso{
\code{\link{regressionComparison}} \code{\link{generateMarginalEffect}} \code{\link{plotMarginalEffect}} \code{\link{plotComplierATE}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
