% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMarginalEffect.R
\name{generateMarginalEffect}
\alias{generateMarginalEffect}
\title{generateMarginalEffect}
\usage{
generateMarginalEffect(temp_model)
}
\arguments{
\item{dataframe}{Dataframe that contains the prompt and response of the open-ended manipulation check.}
}
\value{
Vector of similarity measures (same length as input dataframe)
}
\description{
Calculate the overall treatment effects differing how inattentive participants are down-weighted.
}
\examples{
generateMarginalEffect(temp_model)

}
\seealso{
\code{\link{regressionComparison}} \code{\link{executeMarginalEffect}} \code{\link{plotMarginalEffect}} \code{\link{plotComplierATE}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
