% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMarginalEffect.R
\name{generateMarginalEffect}
\alias{generateMarginalEffect}
\title{Generate Marginal Effects}
\usage{
generateMarginalEffect(unique_covars, simulated_betas, diff_labs,
  model_type, plotDifferences)
}
\arguments{
\item{unique_covars}{Model matrix of unique characteristics used to generate treatment effects.}

\item{simulated_betas}{−XBetas that have been simulated from mvrnorm distribution}

\item{diff_labs}{}

\item{model_type}{Statistical model to estimate. Currently support OLS and logistic ("ls", "logit").}

\item{plotDifferences}{Do you want to see the marginal effects by model, or the differences between the models with regard to their marginal effects? Default=FALSE.}
}
\value{
Dataframe of marginal effects with corresponding 95% confidence intervals.
}
\description{
Internal function that helps calculate the overall treatment effects differing how inattentive participants are down-weighted.
}
\examples{
generateMarginalEffect(generateMarginalEffect(unique_covars = unique_dummies, 
simulated_betas=sim_betas, diff_labs=fd_labs[,1]))

}
\seealso{
\code{\link{regressionComparison}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
