% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimilarity.R
\name{plotSimilarity}
\alias{plotSimilarity}
\title{plotSimilarity}
\usage{
plotSimilarity(dataframe, measure)
}
\arguments{
\item{dataframe}{Dataframe that contains the similarity measures you want to plot.}

\item{measure}{Vector in dataframe that contains the similarity measures you want to plot.}
}
\description{
This function plots the raw distribution for a given similarity measure.
}
\examples{

plotSimilarity(dataframe=replication_complete.cases, measure="jaccardDist")

}
\seealso{
\code{\link{similarity_measures}}
}
\author{
Jeffrey Ziegler
}
