% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimilarity.R
\name{plot_similarity}
\alias{plot_similarity}
\title{plotSimilarity}
\usage{
plot_similarity(dataframe, method = "jaccard")
}
\description{
Plot similarity measure
}
\details{
This function plots the raw distribution for a given similarity measure
}
\examples{

\dontrun{
plotSimilarity(dataframe=kaneData, method="jaccard")
}
}
\seealso{
\code{\link{similarity_measures}}
}
\author{
Jeffrey Ziegler
}
