% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComplierATE.R
\name{plotComplierATE}
\alias{plotComplierATE}
\title{Plot the ATE for Compliers and Non-Compliers}
\usage{
plotComplierATE(dataframe = NULL, similarity_measures = NULL,
  bounds = NULL, n = 100, seed = 12345, formulas = NULL,
  model_type = NULL, k = NULL)
}
\arguments{
\item{dataframe}{Dataframe from which we will estimate our regression model.}

\item{similarity_measures}{Vector(s) from dataframe that contains the similarity measures to be used as weights. Possible values for measure_type = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"). Default is "jaccard".}

\item{bounds}{Minimum and maximum of uniform distribution we should draw cutoff values between.}

\item{n}{Number of simulation rounds/iterations.}

\item{model_type}{Statistical model to estimate. Currently support OLS and logistic ("ls", "logit").}

\item{k}{The penalty that you want to set for down-weighting inattentive respondents. Lower levels of k down-weight low attention participants more severely.}

\item{formula}{Symbolic representation of the model to be estimated. This is written in "typical" R language (i.e. y ~ x1 + x2), such that y is the outcome variable and x1 and x2 are the predictors.}
}
\value{
Plot of the marginal effects for "compliers" and "non-compliers".
}
\description{
Plot the marginal effects for respondents that likely received and did not receive the treatment.
}
\examples{
plotComplierATE(dataframe=replication_complete.cases, similarity_measures=c("jaccardDist", "cosineDist"), bounds=c(0.05, 0.2), n=2, seed=12345, k=3, formula=list(trustLM, responsiveLM), model_type="ls")
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
