% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAttention.R
\name{predictAttention}
\alias{predictAttention}
\title{Predict Attention Among Respondents}
\usage{
predictAttention(
  dataframe = NULL,
  attention_formula = NULL,
  similarity_measures = c("jaccardSimilarity", "cosineSimilarity"),
  correct_vec = NULL,
  model_type = NULL,
  k = 3,
  up_down_weight = "down"
)
}
\arguments{
\item{dataframe}{Dataframe from which you will select attention, predictors for attention, etc. for the regression model.}

\item{k}{The penalty that you want to set for down-weighting inattentive respondents. Lower levels of k down-weight low attention participants more severely.}

\item{up_down_weight}{Do you want to up-weight or down-weight? Default="down".}

\item{attention_predictors}{Symbolic representation of the model to be estimated. This is written in "typical" R language (i.e. x1 + x2). Do not need to specify y or outcome variable, it's assigned internally.}
}
\value{
Object containing instrumental variable regression call, data, results, etc.
}
\description{
Estimate a linear regression in which attention is the outcome, and the user defines the characteristics used to predict attention.
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
