% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionComparison.R
\name{regressionComparison}
\alias{regressionComparison}
\title{Estimate regression models for comparison}
\usage{
regressionComparison(dataframe, formula, similarity_measures, k, model_type)
}
\arguments{
\item{dataframe}{Dataframe from which you will select the outcome, predictors, and weights for the regression model.}

\item{formula}{Symbolic representation of the model to be estimated. This is written in "typical" R language (i.e. y ~ x1 + x2), such that y is the outcome variable and x1 and x2 are the predictors.}

\item{similarity_measures}{Vector(s) from dataframe that contains the similarity measures to be used as weights.}

\item{k}{The penalty that you want to set for down-weighting inattentive respondents. Lower levels of k down-weight low attention participants more severely.}

\item{model_type}{Statistical model to estimate. Currently support OLS and logistic ("ls", "logit").}
}
\value{
Three regression objects are estimated and loaded to your global environment as separate Zelig objects. They are named baseModel, listwiseModel, and weightedModel.
}
\description{
Compare the overall results varying on which respondents are included in the model based on attention.
}
\details{
This function generates three regressions with:
(1) the full sample irrespective of attention
(2) a subsetted sample based on list-wise deletion of respondents that "failed" manipulation check
(3) weighted sample by formula outlined in Ziegler (2020, 5) for average similarity.
}
\examples{
regressionComparison(formula=trustChurch_postTreat ~ Concordant*attendanceBin, dataframe=replication_complete.cases, k=3, similarity_measures=c("jaccardDist", "cosineDist"), model_type="ls")

}
\seealso{
\code{\link{plotMarginalEffect}} \code{\link{plotComplierATE}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
