% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityMeasures.R
\name{similarityMeasures}
\alias{similarityMeasures}
\title{Calculate Similarity Measures}
\usage{
similarityMeasures(dataframe, prompt, response,
  similarity_measures_to_calculate = c("jaccard"), ngrams = 3)
}
\arguments{
\item{dataframe}{Dataframe that contains the prompt and response of the open-ended manipulation check.}

\item{prompt}{Vector/column within `dataframe` that contains the prompt.}

\item{response}{Vector/column within `dataframe` that contains the open-ended response.}

\item{similarity_measures_to_calculate}{Vector of similarity measures to calculate. Possible values for measure_type = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"). Default is "jaccard".}

\item{ngrams}{The number of grams/segments words should be broken into. Default n=3.}
}
\value{
Original dataframe plus columns with similarity measures (same length as input dataframe)
}
\description{
Calculate individual or multiple measures of similarity between prompt and response in open-ended manipulation check.
}
\examples{
replication_complete.cases <- simlarityMeasures(replication_complete.cases, prompt="textViewed", response="validityCheck", ngrams=3)

}
\seealso{
\code{\link{plotSimilarity}} \code{\link{plotSimilarityCorr}}
}
\author{
Jeffrey Ziegler (<jeffrey.ziegler[at]emory.edu>)
}
